+++ instructor-service/src/main/java/net/etg/instructorservice/service/impl/InstructorServiceImpl.java
@@ -5,6 +5,7 @@ import net.etg.instructorservice.dto.CourseDTO;
 import net.etg.instructorservice.dto.InstructorDTO;
 import net.etg.instructorservice.dto.StudentDTO;
 import net.etg.instructorservice.entity.Instructor;
+import net.etg.instructorservice.exception.NoInstructorException;
 import net.etg.instructorservice.proxy.OpenFeignCourseProxy;
 import net.etg.instructorservice.proxy.OpenFeignRegisterProxy;
 import net.etg.instructorservice.repository.InstructorRepository;
@@ -28,38 +29,38 @@ public class InstructorServiceImpl implements InstructorService {
     public InstructorDTO saveInstructor(InstructorDTO instructorDTO){
         Instructor instructor = modelMapper.map(instructorDTO,Instructor.class);
         Instructor savedInstructor = instructorRepository.save(instructor);
-        InstructorDTO savedInstructDTO = modelMapper.map(savedInstructor, InstructorDTO.class);
-        return savedInstructDTO;
+        return modelMapper.map(savedInstructor, InstructorDTO.class);
     }
 
 
     public InstructorDTO getInstructor(String instructorId){
-        Instructor existingInstructor = instructorRepository.findById(instructorId).get();
-        InstructorDTO existingInstructorDTO = modelMapper.map(existingInstructor,InstructorDTO.class);
-        return existingInstructorDTO;
+        Instructor existingInstructor = instructorRepository.findById(instructorId).orElseThrow(NoInstructorException::new);
+        return modelMapper.map(existingInstructor,InstructorDTO.class);
     }
 
     public InstructorDTO updateInstructor(String instructorId,InstructorDTO instructorDTO){
-        Instructor existingInstructor = instructorRepository.findById(instructorId).get();
+        Instructor existingInstructor = instructorRepository.findById(instructorId).orElseThrow(NoInstructorException::new);
         existingInstructor.setName(instructorDTO.getName());
         existingInstructor.setLastName(instructorDTO.getLastName());
         Instructor updatedInstructor = instructorRepository.save(existingInstructor);
-        InstructorDTO updatedInstructorDTO = modelMapper.map(updatedInstructor,InstructorDTO.class);
-        return updatedInstructorDTO;
+        return modelMapper.map(updatedInstructor,InstructorDTO.class);
     }
 
     public String deleteInstructor(String instructorId){
-        instructorRepository.deleteById(instructorId);
-        return "Instructor with code " + instructorId + " deleted";
+        if(instructorRepository.findById(instructorId).isPresent()){
+            instructorRepository.deleteById(instructorId);
+            return "Instructor with code " + instructorId + " deleted";
+        }else{
+            throw new NoInstructorException();
+        }
     }
 
 
     public CourseDTO createACourse(CourseDTO courseDTO){
-        Instructor instructor = instructorRepository.findById(courseDTO.getInstructorId()).get();
+        Instructor instructor = instructorRepository.findById(courseDTO.getInstructorId()).orElseThrow(NoInstructorException::new);
         instructor.addCourse(courseDTO.getCourseCode());
         instructorRepository.save(instructor);
-        CourseDTO savedCourseDTO = openFeignCourseProxy.saveCourse(courseDTO).getBody();
-        return savedCourseDTO;
+        return openFeignCourseProxy.saveCourse(courseDTO).getBody();
     }
 
 
@@ -69,8 +70,7 @@ public class InstructorServiceImpl implements InstructorService {
 
     @Override
     public List<StudentDTO> bringEnrolledStudents(String instructorId, String courseCode) {
-        List<StudentDTO> studentDTOList = openFeignRegisterProxy.getStudents(courseCode).getBody();
-        return studentDTOList;
+        return openFeignRegisterProxy.getStudents(courseCode).getBody();
     }
 
 }
